+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
     - Add a timeline index column to the bookmarks table
     - Add a thread timeline index column to the multi-thread timeline
+    - New same-coord and custom-same-coord power steppers
     
 ### Changes